Extract `SitesWithThisPluginp` and `SitesWithoutThisPlugin` components
