+++ client/dashboard/plugins/plugin/index.tsx
@@ -1,11 +1,8 @@
 import { Site } from '@automattic/api-core';
 import { wpOrgPluginQuery, pluginsQuery, sitesQuery } from '@automattic/api-queries';
 import { useQuery } from '@tanstack/react-query';
-import { Link } from '@tanstack/react-router';
 import { __experimentalVStack as VStack } from '@wordpress/components';
-import { DataViews, filterSortAndPaginate, View } from '@wordpress/dataviews';
 import { __, _n, sprintf } from '@wordpress/i18n';
-import { useMemo, useState } from 'react';
 import { useLocale } from '../../app/locale';
 import { pluginRoute } from '../../app/router/plugins';
 import { DataViewsCard } from '../../components/dataviews-card';
@@ -14,158 +11,8 @@ import PageLayout from '../../components/page-layout';
 import { SectionHeader } from '../../components/section-header';
 import { Text } from '../../components/text';
 import { TextBlur } from '../../components/text-blur';
-
-const defaultSitesWithThisPluginView: View = {
-	type: 'table',
-	fields: [ 'activate', 'autoupdate', 'update' ],
-	sort: { field: 'name', direction: 'asc' },
-	titleField: 'domain',
-};
-
-const SitesWithThisPlugin = ( { sitesWithThisPlugin }: { sitesWithThisPlugin: Site[] } ) => {
-	const [ view, setView ] = useState< View >( defaultSitesWithThisPluginView );
-	const { pluginId } = pluginRoute.useParams();
-	const locale = useLocale();
-	const { data: sitesPlugins, isLoading: isLoadingSitesPlugins } = useQuery( pluginsQuery() );
-	const { isLoading: isLoadingSites } = useQuery( sitesQuery() );
-	const { isLoading: isLoadingWpOrgPlugin } = useQuery( wpOrgPluginQuery( pluginId, locale ) );
-
-	const fields = useMemo(
-		() => [
-			{
-				id: 'domain',
-				label: __( 'Site' ),
-				getValue: ( { item }: { item: Site } ) => item.URL,
-				render: ( { item }: { item: Site } ) => item.URL,
-				enableHiding: false,
-				enableSorting: true,
-				enableGlobalSearch: true,
-			},
-			{
-				id: 'active',
-				label: __( 'Active' ),
-				getValue: ( { item }: { item: Site } ) =>
-					sitesPlugins?.sites[ item.ID ].find( ( p ) => p.slug === pluginId )?.active ?? false,
-				render: ( { item }: { item: Site } ) =>
-					sitesPlugins?.sites[ item.ID ].find( ( p ) => p.slug === pluginId )?.active ?? false,
-				enableHiding: false,
-				enableSorting: true,
-			},
-			{
-				id: 'autoupdate',
-				label: __( 'Autoupdate' ),
-				getValue: ( { item }: { item: Site } ) =>
-					sitesPlugins?.sites[ item.ID ].find( ( p ) => p.slug === pluginId )?.autoupdate ?? false,
-				render: ( { item }: { item: Site } ) =>
-					sitesPlugins?.sites[ item.ID ].find( ( p ) => p.slug === pluginId )?.autoupdate ?? false,
-				enableHiding: false,
-				enableSorting: true,
-			},
-			{
-				id: 'update',
-				label: __( 'Update' ),
-				render: () => 'Update',
-				enableHiding: false,
-				enableSorting: false,
-			},
-		],
-		[ pluginId, sitesPlugins ]
-	);
-
-	const { data, paginationInfo } = filterSortAndPaginate( sitesWithThisPlugin, view, fields );
-
-	return (
-		<DataViews
-			isLoading={ isLoadingSitesPlugins || isLoadingSites || isLoadingWpOrgPlugin }
-			data={ data }
-			fields={ fields }
-			view={ view }
-			onChangeView={ setView }
-			defaultLayouts={ { table: {} } }
-			actions={ [
-				{
-					id: 'delete',
-					label: __( 'Delete' ),
-					isPrimary: false,
-					callback: ( items ) => {
-						// Dummy delete action for now
-						// eslint-disable-next-line no-console
-						console.log( 'Delete clicked for plugin', items[ 0 ] );
-					},
-				},
-			] }
-			getItemId={ ( item ) => String( item.ID ) }
-			paginationInfo={ paginationInfo }
-		/>
-	);
-};
-
-const defaultSitesWithoutThisPluginView: View = {
-	type: 'table',
-	fields: [ 'link' ],
-	layout: {
-		styles: {
-			link: { align: 'end' },
-		},
-	},
-	sort: { field: 'name', direction: 'asc' },
-	titleField: 'domain',
-};
-
-const SitesWithoutThisPlugin = ( {
-	sitesWithoutThisPlugin,
-}: {
-	sitesWithoutThisPlugin: Site[];
-} ) => {
-	const [ view, setView ] = useState< View >( defaultSitesWithoutThisPluginView );
-	const { pluginId } = pluginRoute.useParams();
-	const locale = useLocale();
-	const { isLoading: isLoadingSitesPlugins } = useQuery( pluginsQuery() );
-	const { isLoading: isLoadingSites } = useQuery( sitesQuery() );
-	const { isLoading: isLoadingWpOrgPlugin } = useQuery( wpOrgPluginQuery( pluginId, locale ) );
-
-	const fields = useMemo(
-		() => [
-			{
-				id: 'domain',
-				label: __( 'Site' ),
-				getValue: ( { item }: { item: Site } ) => item.URL,
-				render: ( { item }: { item: Site } ) => item.URL,
-				enableHiding: false,
-				enableSorting: true,
-				enableGlobalSearch: true,
-			},
-			{
-				id: 'link',
-				header: <div />,
-				getValue: ( { item }: { item: Site } ) => item.URL,
-				render: ( { item }: { item: Site } ) => (
-					<Link to={ `/plugins/${ pluginId }/${ item.slug }` }>{ __( 'Go to plugin page' ) }</Link>
-				),
-			},
-		],
-		[ pluginId ]
-	);
-
-	const { data, paginationInfo } = filterSortAndPaginate( sitesWithoutThisPlugin, view, fields );
-
-	return (
-		<DataViews
-			search={ false }
-			isLoading={ isLoadingSitesPlugins || isLoadingSites || isLoadingWpOrgPlugin }
-			data={ data }
-			fields={ fields }
-			view={ view }
-			onChangeView={ setView }
-			defaultLayouts={ { table: {} } }
-			getItemId={ ( item ) => String( item.ID ) }
-			paginationInfo={ paginationInfo }
-		>
-			<DataViews.Layout />
-			<DataViews.Pagination />
-		</DataViews>
-	);
-};
+import { SitesWithThisPlugin } from './sites-with-this-plugin';
+import { SitesWithoutThisPlugin } from './sites-without-this-plugin';
 
 export default function Plugin() {
 	const { pluginId } = pluginRoute.useParams();
@@ -241,7 +88,7 @@ export default function Plugin() {
 					/>
 
 					<DataViewsCard>
-						<SitesWithThisPlugin sitesWithThisPlugin={ sitesWithThisPlugin } />
+						<SitesWithThisPlugin sites={ sitesWithThisPlugin } />
 					</DataViewsCard>
 				</VStack>
 
@@ -249,7 +96,7 @@ export default function Plugin() {
 					<SectionHeader title={ __( 'Available on' ) } />
 
 					<DataViewsCard>
-						<SitesWithoutThisPlugin sitesWithoutThisPlugin={ sitesWithoutThisPlugin } />
+						<SitesWithoutThisPlugin sites={ sitesWithoutThisPlugin } />
 					</DataViewsCard>
 				</VStack>
 			</VStack>
